Added smooth transition for back to top button
I have added the smooth transition button which takes you to the top of the website. There are some hover effects also and it appears when the screen size is greater than 100. I have added back to top button to the pages like home, blog, quiz, tools and login sections.